Fix float restore_timeout breaking IPMI monitor script
CONF.playbook_timeout * 1.5 produces a float which reaches
ipmi_sel_event_monitor.py's --timeout argument, causing argparse
to fail with "invalid int value: '5400.0'". This commit casts
the restore_timeout to int.

Test Plan:
01. PASS - Trigger a subcloud factory restore without specifying
    restore_timeout. Verify the IPMI monitoring script starts without
    "invalid int value". Repeat the test with the auto-restore
    operation.
02. PASS - Trigger a subcloud auto restore with an explicit
    restore_timeout. Verify it is passed through correctly.

Closes-bug: 2144352

+++ distributedcloud/dcmanager/manager/subcloud_manager.py
@@ -2857,8 +2857,10 @@ class SubcloudManager(manager.Manager):
             # The restore timeout needs to be increased a half of the default
             # global ansible playbook timeout because the default of 1h is not
             # enough to restore a duplex subcloud running rook ceph.
-            restore_timeout = payload.get("restore_values", {}).get(
-                "restore_timeout", CONF.playbook_timeout * 1.5
+            restore_timeout = int(
+                payload.get("restore_values", {}).get(
+                    "restore_timeout", CONF.playbook_timeout * 1.5
+                )
             )
 
             overrides_file = self._create_overrides_for_backup_or_restore(
